add foreign key for posts and fix some methods

+++ app/routers/post.py
@@ -14,10 +14,19 @@ router = APIRouter(
 @router.get(
     "/",
     response_model=list[schemas.Post])
-def get_posts(db: Session = Depends(get_db)):
+def get_posts(
+    db: Session = Depends(get_db),
+    current_user: int = Depends(oauth2.get_current_user),
+    limit: int = 10,
+    skip: int = 0,
+    search: str | None = ""
+):
     posts = (
         db
         .query(models.Post)
+        .filter(models.Post.title.contains(search))
+        .limit(limit)
+        .offset(skip)
         .all()
     )
     return posts
@@ -32,7 +41,7 @@ def create_posts(
     db: Session = Depends(get_db),
     current_user: int = Depends(oauth2.get_current_user)
 ):
-    new_post = models.Post(**post.dict())
+    new_post = models.Post(owner_id=current_user.id, **post.dict())
 
     db.add(new_post)
     db.commit()
@@ -54,7 +63,6 @@ def get_post(
         .filter(models.Post.id == id)
         .first()
     )
-    print(post)
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
 
@@ -78,6 +86,9 @@ def delete_post(
 
     if post is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
+
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
 
     post_query.delete(synchronize_session=False)
     db.commit()
@@ -103,6 +114,9 @@ def update_post(
 
     if post is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
+
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
 
     post_query.update(updated_post.dict(), synchronize_session=False)
     db.commit()
